Agregue metodos en controlador de especies

+++ src/main/java/com/api/pnat/interfaces/IAvistamiento.java
@@ -1,10 +1,7 @@
 package com.api.pnat.interfaces;
 
-
 import com.api.pnat.model.Avistamiento;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IAvistamiento extends CrudRepository<Avistamiento, Long> {
-
-//    public Empresa findByIdEmpresaAndNombre(int id, String nombre);
+public interface IAvistamiento extends CrudRepository<Avistamiento, Integer> {
 }
