+++ src/main/java/com/api/pnat/interfaces/IEspecies.java
@@ -1,10 +1,9 @@
 package com.api.pnat.interfaces;
 
-import com.api.pnat.model.Avistamiento;
 import com.api.pnat.model.Especies;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface IEspecies extends CrudRepository<Especies, Long> {
+public interface IEspecies extends CrudRepository<Especies, Integer> {
 
 }
